Cooking with OOP finally.

+++ include/string.cpp
@@ -8,7 +8,7 @@
 
 
 
-int strlen(const char* str)
+size_t strlen(const char* str)
 {
     int len = 0;
     while (str[len])
