+++ include/string.h
@@ -5,7 +5,7 @@
 #ifndef STRING_H
 #define STRING_H
 
-#include "serial.h"
+
 #include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
@@ -15,7 +15,7 @@ static constexpr char hex[16] = {'0', '1', '2', '3', '4', '5', '6', '7', '8', '9
 
 
 // extern "C"
-int strlen(const char* str);
+size_t strlen(const char* str);
 
 //extern "C"
 template <typename int_like>
